Better test separation and naming
